requirements.txt
cryptography==45.0.5
django-ckeditor==6.7.3
django-grappelli==4.0.2
django-js-asset==3.1.2
mysql==0.0.3
mysqlclient==2.2.7
pillow==11.2.1
pip==25.1.1
pycparser==2.22
PyMySQL==1.1.1
sqlparse==0.5.3
tzdata==2025.2